+++ packages/vue/src/components/sign-up.vue
@@ -1,52 +1,57 @@
 <script setup lang="ts">
 import { computed, useAttrs, toRefs } from 'vue';
+
 import {
   authenticatorTextUtil,
   getFormDataFromEvent,
   translate,
 } from '@aws-amplify/ui';
 
-import FederatedSignIn from './federated-sign-in.vue';
-import AuthenticatorSignUpFormFields from './authenticator-sign-up-form-fields.vue';
-
 import { useAuthenticator } from '../composables/useAuth';
+import { UseAuthenticator } from '../types';
+import AuthenticatorSignUpFormFields from './authenticator-sign-up-form-fields.vue';
+import FederatedSignIn from './federated-sign-in.vue';
 
-const props = useAuthenticator();
-const { hasValidationErrors, isPending, error } = toRefs(props);
+// `facade` is manually typed to `UseAuthenticator` for temporary type safety.
+const facade: UseAuthenticator = useAuthenticator();
+const { submitForm, updateBlur, updateForm } = facade;
+const { error, hasValidationErrors, isPending } = toRefs(facade);
 
 const attrs = useAttrs();
+
+/** @deprecated Authenticator component events are deprecated and not maintained. */
 const emit = defineEmits(['signUpSubmit']);
 
 // Text Util
 const { getCreateAccountText } = authenticatorTextUtil;
+
 // Computed Properties
 const createAccountLabel = computed(() => getCreateAccountText());
 
 // Methods
-
 const onInput = (e: Event): void => {
   let { checked, name, type, value } = e.target as HTMLInputElement;
 
-  if (type === 'checkbox' && !checked)
-    (value as string | undefined) = undefined;
-  props.updateForm({ name, value });
+  const isUncheckedCheckbox = type === 'checkbox' && !checked;
+  updateForm({
+    name,
+    value: isUncheckedCheckbox ? undefined : (value as string),
+  });
 };
 
 function onBlur(e: Event) {
   const { name } = e.target as HTMLInputElement;
-  props.updateBlur({ name });
+  updateBlur({ name });
 }
 
 const onSignUpSubmit = (e: Event): void => {
+  // TODO(BREAKING): remove unused emit
+  // istanbul ignore next
   if (attrs?.onSignUpSubmit) {
     emit('signUpSubmit', e);
   } else {
-    submit(e);
+    submitForm(getFormDataFromEvent(e));
   }
-};
-
-const submit = (e: Event): void => {
-  props.submitForm(getFormDataFromEvent(e));
 };
 </script>
 
@@ -82,8 +87,9 @@ const submit = (e: Event): void => {
             :variation="'primary'"
             style="border-radius: 0px; font-weight: normal"
             :disabled="isPending || hasValidationErrors"
-            >{{ createAccountLabel }}</amplify-button
           >
+            {{ createAccountLabel }}
+          </amplify-button>
         </base-wrapper>
       </base-form>
     </base-wrapper>